rename "internal" to "lib"
Two "internal" folders were nested,
renamed one of them to "lib" as it maded more sense.
