+++ app/safe/internal/server/route/lib/handle/response.go
