+++ app/safe/internal/server/route/list/impl.go
@@ -16,7 +16,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/vmware-tanzu/secrets-manager/app/safe/internal/server/route/internal/validation"
+	"github.com/vmware-tanzu/secrets-manager/app/safe/internal/server/route/lib/validation"
 	"github.com/vmware-tanzu/secrets-manager/app/safe/internal/state/secret/collection"
 	"github.com/vmware-tanzu/secrets-manager/core/audit/journal"
 	event "github.com/vmware-tanzu/secrets-manager/core/audit/state"
